requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
apache-airflow==2.6.3
apache-airflow-providers-common-sql==1.5.2
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.4.2
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
apturl==0.5.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.4
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
command-not-found==0.3
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==40.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docutils==0.20.1
duplicity==0.8.21
email-validator==1.3.1
exceptiongroup==1.1.2
fasteners==0.14.1
filelock==3.6.0
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
future==0.18.2
google-re2==1.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
idna==3.4
importlib-metadata==4.6.4
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
packaging==21.3
pandas==2.0.3
paramiko==2.9.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
pluggy==1.2.0
prison==0.2.1
protobuf==3.12.4
psutil==5.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.11
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
pyxdg==0.27
PyYAML==6.0
referencing==0.29.1
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.4.2
rich-argparse==1.2.0
rpds-py==0.8.8
SecretStorage==3.3.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.16
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
